Move pdf_file to models.py
